refactor: 게임 상태 관련 class gamestatus 패키지로 이동
