+++ src/main/java/janggi/service/JanggiService.java
@@ -5,9 +5,9 @@ import janggi.dao.PieceDao;
 import janggi.dao.entity.GameEntity;
 import janggi.dao.entity.PieceEntity;
 import janggi.dao.entity.Status;
-import janggi.domain.GameStatus;
 import janggi.domain.board.BoardSetUp;
 import janggi.domain.board.JanggiBoard;
+import janggi.domain.gamestatus.GameStatus;
 import janggi.domain.piece.Cannon;
 import janggi.domain.piece.Chariot;
 import janggi.domain.piece.ChuSoldier;
