+++ src/test/java/janggi/service/JanggiServiceTest.java
@@ -9,12 +9,12 @@ import janggi.dao.PieceDao;
 import janggi.dao.entity.GameEntity;
 import janggi.dao.entity.PieceEntity;
 import janggi.dao.entity.Status;
-import janggi.domain.GameEnded;
-import janggi.domain.GameRunned;
-import janggi.domain.GameStatus;
 import janggi.domain.board.ChuBoardSetUp;
 import janggi.domain.board.HanBoardSetUp;
 import janggi.domain.board.JanggiBoard;
+import janggi.domain.gamestatus.GameEnded;
+import janggi.domain.gamestatus.GameRunned;
+import janggi.domain.gamestatus.GameStatus;
 import janggi.domain.piece.Chariot;
 import janggi.domain.piece.Dynasty;
 import janggi.domain.piece.General;
